implement kendo ui

+++ src/app/firebase/user-repository/user-credential-repository/user-credential-repository.service.ts
@@ -1,5 +1,3 @@
-import { Injectable } from '@angular/core';
-import { AngularFirestore } from '@angular/fire/compat/firestore';
 import {
   ChangeNumberUserCriteriaType,
   IUser,
@@ -7,21 +5,39 @@ import {
   UserAssociatedShopType,
   UserSettingType,
 } from 'src/app/interface';
+import { Injectable, inject } from '@angular/core';
 import * as Db from 'src/app/constant/firebase-path';
-import { Observable, catchError, firstValueFrom, from, map, of, switchMap, take } from 'rxjs';
+import { Observable, firstValueFrom, from, map, of, switchMap, take } from 'rxjs';
 import { FirebaseToasterService } from '../../firebase-toaster/firebase-toaster.service';
 import { override } from 'functions/src/service/override/user/user-setting-override/user-setting-override';
 import { SystemLanguageStorageService } from 'src/app/service/global/language/system-language-management/system-language-storage/system-language-storage.service';
 import { TextTransformService } from 'src/app/service/global/text-transform/text-transform.service';
 import * as Constant from 'src/app/constant/constant';
+import { FirebaseApiService, createKeyMap, Query } from '../../firebase-api/firebase-api.service';
+
+const userParam = createKeyMap<IUser>([
+  'id',
+  'firstName',
+  'lastName',
+  'associatedShopIds',
+  'disabledAccount',
+  'dob',
+  'email',
+  'gender',
+  'isSystemAdmin',
+  'phoneNumber',
+  'signature',
+  'visitedShopIds',
+]);
+
 @Injectable({
   providedIn: 'root',
 })
 export class UserCredentialRepositoryService {
+  private _api = inject(FirebaseApiService);
   private readonly _timeStamp = { lastModifiedDate: new Date() };
   private readonly _emailRegex = /^[a-zA-Z0-9._-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,6}$/;
   constructor(
-    private _afs: AngularFirestore,
     private _toaster: FirebaseToasterService,
     private _textTransform: TextTransformService,
     private _systemLanguageStorage: SystemLanguageStorageService
@@ -46,125 +62,117 @@ export class UserCredentialRepositoryService {
   }
 
   public subscribeUserById(userId: string): Observable<IUser | null> {
-    const userCollectionRef = this._afs.collection<IUser>(Db.Context.User, ref =>
-      ref.where('id', '==', userId).limit(1)
-    );
-
-    return userCollectionRef.get().pipe(
-      map(userQuerySnapshot => {
-        if (userQuerySnapshot.docs.length > 0) {
-          const user = userQuerySnapshot.docs[0].data();
-          user.setting = this.overrideSetting(user.setting);
-          return user;
-        } else {
-          return null;
-        }
-      })
-    );
+    return this._api
+      .getDocument<IUser>(Db.Context.User, ref => ref.where(userParam.id, Query.Equal, userId).limit(1))
+      .pipe(
+        map(user => {
+          if (user !== null) {
+            user.setting = this.overrideSetting(user.setting);
+            return user;
+          } else {
+            return null;
+          }
+        })
+      );
   }
 
   public subscribeUserByPhoneNumber(phoneNumber: string): Observable<IUser | null> {
-    const userCollectionRef = this._afs.collection<IUser>(Db.Context.User, ref =>
-      ref.where('phoneNumber', '==', phoneNumber).where('loginOption.phoneNumber', '==', true).limit(1)
-    );
-
-    return userCollectionRef.get().pipe(
-      map(userQuerySnapshot => {
-        if (userQuerySnapshot.docs.length > 0) {
-          const user = userQuerySnapshot.docs[0].data();
-          user.setting = this.overrideSetting(user.setting);
-          return user;
-        } else {
-          return null;
-        }
-      })
-    );
+    return this._api
+      .getDocument<IUser>(Db.Context.User, ref =>
+        ref
+          .where(userParam.phoneNumber, Query.Equal, phoneNumber)
+          .where('loginOption.phoneNumber', '==', true)
+          .limit(1)
+      )
+      .pipe(
+        map(user => {
+          if (user !== null) {
+            user.setting = this.overrideSetting(user.setting);
+            return user;
+          } else {
+            return null;
+          }
+        })
+      );
   }
 
   public findIdAndEmailByChangeNumberUserCriteria(criteria: ChangeNumberUserCriteriaType) {
-    const formattedFirstName = this._textTransform.getTitleFormat(criteria.firstName);
-    const formattedLastName = this._textTransform.getTitleFormat(criteria.lastName);
-    const userCollectionRef = this._afs.collection<IUser>(Db.Context.User, ref =>
-      ref
-        .where('firstName', '==', formattedFirstName)
-        .where('lastName', '==', formattedLastName)
-        .where('phoneNumber', '==', criteria.previousPhoneNumber)
-        .where('gender', '==', criteria.gender)
-        .where('email', '==', criteria.email)
-        .where('dob', '==', criteria.dob)
-        .where('loginOption.phoneNumber', '==', true)
-        .limit(1)
-    );
-
-    return userCollectionRef.get().pipe(
-      map(userQuerySnapshot => {
-        if (userQuerySnapshot.docs.length > 0) {
-          const user = userQuerySnapshot.docs[0].data();
-          user.setting = this.overrideSetting(user.setting);
-          return { email: user.email, id: user.id };
-        } else {
-          return null;
-        }
-      })
-    );
+    return this._api
+      .getDocument<IUser>(Db.Context.User, ref =>
+        ref
+          .where(userParam.firstName, Query.Equal, this._textTransform.getTitleFormat(criteria.firstName))
+          .where(userParam.lastName, Query.Equal, this._textTransform.getTitleFormat(criteria.lastName))
+          .where(userParam.phoneNumber, Query.Equal, criteria.previousPhoneNumber)
+          .where(userParam.gender, Query.Equal, criteria.gender)
+          .where(userParam.email, Query.Equal, criteria.email)
+          .where(userParam.dob, Query.Equal, criteria.dob)
+          .where('loginOption.phoneNumber', '==', true)
+          .limit(1)
+      )
+      .pipe(
+        map(doc => {
+          if (doc !== null) {
+            doc.setting = this.overrideSetting(doc.setting);
+            return { email: doc.email, id: doc.id };
+          } else {
+            return null;
+          }
+        })
+      );
   }
 
   public subscribeValueChangeListener(id: string): Observable<IUser | null> {
-    const userDocRef = this._afs.doc<IUser>(`${Db.Context.User}/${id}`);
-
-    return userDocRef.valueChanges().pipe(
-      switchMap(userData => {
-        if (userData) {
-          userData.setting = this.overrideSetting(userData.setting);
-          return of(userData);
-        } else {
-          return of(null);
-        }
-      })
-    );
+    return this._api
+      .getDocument<IUser>(Db.Context.User, ref => ref.where(userParam.id, Query.Equal, id).limit(1))
+      .pipe(
+        map(doc => {
+          if (doc !== null) {
+            doc.setting = this.overrideSetting(doc.setting);
+            return doc;
+          } else {
+            return null;
+          }
+        })
+      );
   }
 
   public subscribeUserByEmail(email: string): Observable<IUser | null> {
-    const userCollectionRef = this._afs.collection<IUser>(Db.Context.User, ref =>
-      ref.where('email', '==', email).where('loginOption.email', '==', true).limit(1)
-    );
-
-    return userCollectionRef.get().pipe(
-      map(userQuerySnapshot => {
-        if (userQuerySnapshot.docs.length > 0) {
-          const user = userQuerySnapshot.docs[0].data();
-          user.setting = this.overrideSetting(user.setting);
-          return user;
-        } else {
-          return null;
-        }
-      }),
-      catchError(error => {
-        throw error;
-      })
-    );
+    return this._api
+      .getDocument<IUser>(Db.Context.User, ref =>
+        ref.where(userParam.email, Query.Equal, email).where('loginOption.email', Query.Equal, true).limit(1)
+      )
+      .pipe(
+        map(doc => {
+          if (doc !== null) {
+            doc.setting = this.overrideSetting(doc.setting);
+            return doc;
+          } else {
+            return null;
+          }
+        })
+      );
   }
 
   public subscribeAssociatedShopUsers(shopId: string): Observable<ShopEmployeeManagementUserType[]> {
-    return this._afs
-      .collection<IUser>(Db.Context.User, ref =>
-        ref.where('associatedShopIds', 'array-contains', shopId).where('isSystemAdmin', '!=', true)
+    return this._api
+      .snapshotChangeDocuments<IUser>(Db.Context.User, ref =>
+        ref
+          .where(userParam.associatedShopIds, Query.ArrayContains, shopId)
+          .where(userParam.isSystemAdmin, Query.NotEqual, true)
       )
-      .snapshotChanges()
       .pipe(
-        map(users =>
-          users.map(user => this.transformIntoShopEmployeeManagementUserType(user.payload.doc.data(), shopId))
-        ),
+        map(users => users.map(user => this.transformIntoShopEmployeeManagementUserType(user, shopId))),
         map(users => users.filter(user => user !== null) as ShopEmployeeManagementUserType[])
       );
   }
 
   public activeShopSpecialist(shopId: string) {
-    return this._afs
-      .collection<IUser>(Db.Context.User, ref =>
-        ref.where('associatedShopIds', 'array-contains', shopId).where('isSystemAdmin', '!=', true)
+    return this._api
+      .valueChangeDocuments<IUser>(Db.Context.User, ref =>
+        ref
+          .where(userParam.associatedShopIds, Query.ArrayContains, shopId)
+          .where(userParam.isSystemAdmin, Query.NotEqual, true)
       )
-      .valueChanges()
       .pipe(
         map(users => users.map(user => this.transformIntoShopEmployeeManagementUserType(user, shopId))),
         map(users => users.filter(user => user !== null) as ShopEmployeeManagementUserType[]),
@@ -174,17 +182,15 @@ export class UserCredentialRepositoryService {
   }
 
   public subscribeAllUser() {
-    return this._afs
-      .collection<IUser>(Db.Context.User, ref => ref.orderBy('isSystemAdmin'))
-      .get()
+    return this._api
+      .getDocuments<IUser>(Db.Context.User, ref => ref.orderBy(userParam.isSystemAdmin))
       .pipe(
-        switchMap(userSnapshots =>
+        switchMap(docs =>
           from(
             Promise.all(
-              userSnapshots.docs.map(snapshot => {
-                let user = snapshot.data();
-                user.setting = this.overrideSetting(user.setting);
-                return user;
+              docs.map(doc => {
+                doc.setting = this.overrideSetting(doc.setting);
+                return doc;
               })
             )
           )
@@ -195,21 +201,28 @@ export class UserCredentialRepositoryService {
   public async createUser(user: IUser): Promise<boolean> {
     let command = { ...user, ...this._timeStamp };
     try {
-      await this._afs.collection(Db.Context.User).doc(user.id).set(command);
-      await this._toaster.addSuccess();
-      return true;
+      const saved = await this._api.set<IUser>(Db.Context.User, command);
+      if (saved) {
+        await this._toaster.addSuccess();
+      }
+      return saved !== null;
     } catch (error) {
       await this._toaster.addFail(error);
       console.error(error);
       return false;
     }
   }
+
   public async updateUser(user: IUser) {
     const updateCriteria = { ...user, ...this._timeStamp };
     try {
-      await this._afs.collection(Db.Context.User).doc(user.id).update(updateCriteria);
-      await this._toaster.updateSuccess();
-      return true;
+      const updated = await this._api.updateDocument<IUser>(Db.Context.User, updateCriteria);
+
+      if (updated) {
+        await this._toaster.updateSuccess();
+      }
+
+      return updated;
     } catch (error) {
       await this._toaster.updateFail(error);
       console.error(error);
@@ -220,8 +233,7 @@ export class UserCredentialRepositoryService {
   public async updateUserWithoutSuccessNotification(user: IUser) {
     const updateCriteria = { ...user, ...this._timeStamp };
     try {
-      await this._afs.collection(Db.Context.User).doc(user.id).update(updateCriteria);
-      return true;
+      return await this._api.updateDocument<IUser>(Db.Context.User, updateCriteria);
     } catch (error) {
       await this._toaster.updateFail(error);
       console.error(error);
@@ -231,9 +243,13 @@ export class UserCredentialRepositoryService {
 
   public async deleteUser(user: IUser) {
     try {
-      await this._afs.collection(Db.Context.User).doc(user.id).delete();
-      await this._toaster.deleteSuccess();
-      return true;
+      const deleted = await this._api.delete<IUser>(Db.Context.User, user.id);
+
+      if (deleted) {
+        await this._toaster.deleteSuccess();
+      }
+
+      return deleted;
     } catch (error) {
       await this._toaster.deleteFail(error);
       console.error(error);
@@ -297,7 +313,7 @@ export class UserCredentialRepositoryService {
 
   public async buildNewUser() {
     const newUser: IUser = {
-      id: this._afs.createId(),
+      id: this._api.newId(),
       firstName: '',
       lastName: '',
       gender: 'Male',
